Dashboard-0.0-main/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.15.0
requests>=2.31.0
openpyxl>=3.1.0
xlrd>=2.0.1
python-dotenv>=1.0.0
gunicorn>=21.0.0
